Add tests for 2 scanners
