+++ src/test/java/com/cx/restclient/general/ScaAndSastScanTests.java
@@ -1,0 +1,57 @@
+package com.cx.restclient.general;
+
+import com.cx.restclient.CxClientDelegator;
+import com.cx.restclient.configuration.CxScanConfig;
+import com.cx.restclient.dto.ScanResults;
+import com.cx.restclient.exception.CxClientException;
+import com.cx.restclient.sca.dto.RemoteRepositoryInfo;
+import com.cx.restclient.sca.dto.SCAResults;
+import com.cx.restclient.sca.dto.SourceLocationType;
+import com.cx.restclient.sca.dto.report.Finding;
+import com.cx.restclient.sca.dto.report.Package;
+import com.cx.restclient.sca.dto.report.SCASummaryResults;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.compress.archivers.ArchiveEntry;
+import org.apache.commons.compress.archivers.ArchiveException;
+import org.apache.commons.compress.archivers.ArchiveInputStream;
+import org.apache.commons.compress.archivers.ArchiveStreamFactory;
+import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.IOUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.junit.Ignore;
+import org.junit.Test;
+
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.List;
+import java.util.UUID;
+
+import static org.junit.Assert.*;
+
+@Slf4j
+public class ScaAndSastScanTests extends AbstractScaScanTests {
+    
+
+    
+    public void scanRemoteRepo(String repoUrlProp, boolean useOnPremAuthentication) throws MalformedURLException {
+        CxScanConfig config = initScaConfig(repoUrlProp, useOnPremAuthentication);
+        config = initSastConfig(config, "SastAndSca");
+        
+        ScanResults scanResults = runScan(config);
+        verifyScanResults(scanResults);
+    }
+
+
+
+
+
+   
+
+}
